Move utils to separate function
